boss scene and credits scene created
